+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+N/A
+
+## [v8.18.0] - 2026-05-31
+
 ### Changed
 
 - `defaultValueProcessor` is now marked `@deprecated` (prefer `defaultValueProcessorByRule`).
@@ -2234,7 +2238,8 @@ _(This list may look long, but the breaking changes should only affect a small m
 
 <!-- #region Release comparison links -->
 
-[unreleased]: https://github.com/react-querybuilder/react-querybuilder/compare/v8.17.0...HEAD
+[unreleased]: https://github.com/react-querybuilder/react-querybuilder/compare/v8.18.0...HEAD
+[v8.18.0]: https://github.com/react-querybuilder/react-querybuilder/compare/v8.17.0...v8.18.0
 [v8.17.0]: https://github.com/react-querybuilder/react-querybuilder/compare/v8.16.2...v8.17.0
 [v8.16.2]: https://github.com/react-querybuilder/react-querybuilder/compare/v8.16.1...v8.16.2
 [v8.16.1]: https://github.com/react-querybuilder/react-querybuilder/compare/v8.16.0...v8.16.1